Split up jipdate.py into several files
There are three major changes here:
 - Global variables are stored in a separate file (avoid using the
   "global" keyword everywhere).
 - The Jira login and instance is in a separate file. That makes it
   possible to re-use this functionality for other scripts in the
   future.
 - For the same reason we moved the vprint and eprint to a separate
   file.
