+++ .gitignore
@@ -1,3 +1,4 @@
 config.yml
 status.txt
 *.swp
+*.pyc
