+++ cfg.py
@@ -1,0 +1,7 @@
+TEST_SERVER = 'https://dev-projects.linaro.org'
+PRODUCTION_SERVER = 'https://projects.linaro.org'
+
+args = None
+config_filename = "config.yml"
+server = PRODUCTION_SERVER
+yml_config = None
